Review comments for refactor
