+++ system_tests/extnet_tests.py
@@ -16,6 +16,7 @@
 import unittest
 from uuid import uuid1
 
+from helpers.portgroup_helper import PortgroupHelper
 from pyvcloud.system_test_framework.base_test import BaseTestCase
 from pyvcloud.system_test_framework.environment import developerModeAware
 from pyvcloud.system_test_framework.environment import Environment
@@ -66,26 +67,13 @@ class TestExtNet(BaseTestCase):
 
         platform = Platform(TestExtNet._sys_admin_client)
         vc_name = TestExtNet._config['vc']['vcenter_host_name']
-        name_filter = ('vcName', vc_name)
-        query = TestExtNet._sys_admin_client.get_typed_query(
-            ResourceType.PORT_GROUP.value,
-            query_result_format=QueryResultFormat.RECORDS,
-            equality_filter=name_filter)
-
-        for record in list(query.execute()):
-            if record.get('networkName') == '--':
-                if record.get('portgroupType') == TestExtNet._portgroupType  \
-                    and not record.get('name').startswith('vxw-'):
-                    TestExtNet._port_group = record.get('name')
-                    break
-
-        self.assertIsNotNone(self._port_group,
-                             'None of the port groups are free.')
+        portgrouphelper = PortgroupHelper(TestExtNet._sys_admin_client)
+        pg_name = portgrouphelper.get_pgroup_name(vc_name, TestExtNet._portgroupType)
 
         ext_net = platform.create_external_network(
             name=TestExtNet._name,
             vim_server_name=vc_name,
-            port_group_names=[TestExtNet._port_group],
+            port_group_names= [pg_name],
             gateway_ip=TestExtNet._gateway,
             netmask=TestExtNet._netmask,
             ip_ranges=[TestExtNet._ip_range],
@@ -276,13 +264,13 @@ class TestExtNet(BaseTestCase):
        """
         logger = Environment.get_default_logger()
         platform = Platform(TestExtNet._sys_admin_client)
-        vim_server_name = TestExtNet._config['vc2']['vcenter_host_name']
+        vc_name = TestExtNet._config['vc2']['vcenter_host_name']
+        portgrouphelper = PortgroupHelper(TestExtNet._sys_admin_client)
+        pg_name = portgrouphelper.get_pgroup_name(vc_name, TestExtNet._portgroupType)
 
-        pgroup_name = platform.get_pgroup_name(vim_server_name,
-            TestExtNet._portgroupType)
         ext_net = self._get_ext_net(platform).attach_port_group(
-            vim_server_name,
-            pgroup_name)
+            vc_name,
+            pg_name)
         task = ext_net['{' + NSMAP['vcloud'] + '}Tasks'].Task[0]
         TestExtNet._sys_admin_client.get_task_monitor().wait_for_success(
         task=task)
@@ -291,7 +279,7 @@ class TestExtNet(BaseTestCase):
         + TestExtNet._name + '.')
         ext_net = platform.get_external_network(self._name)
         self.assertIsNotNone(ext_net)
-        vc_record = platform.get_vcenter(vim_server_name)
+        vc_record = platform.get_vcenter(vc_name)
         vc_href = vc_record.get('href')
         vim_port_group_refs = \
             ext_net['{' + NSMAP['vmext'] + '}VimPortGroupRefs']
